backend/requirements_docker.txt
accelerate==1.8.1
alembic==1.16.2
altair 
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.6.2
asttokens 
async-timeout==5.0.1
attrs 
bcrypt==4.3.0
black==25.1.0
blinker 
Bottleneck 
brotlicffi 
cachetools 
certifi==2025.6.15
cffi 
chardet==5.2.0
charset-normalizer 
click 
comm 
datamodel-code-generator==0.33.0
debugpy 
decorator 
Deprecated 
distro==1.9.0
exceptiongroup 
executing 
filelock 
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec 
genson==1.3.0
gitdb 
GitPython 
gmpy2 
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata 
inflect==7.5.0
ipykernel 
ipython 
isort==6.0.1
itsdangerous==2.2.0
jedi 
Jinja2 
jiter==0.12.0
jsonschema 
jsonschema-specifications 
jupyter_client 
jupyter_core 
lit 
Mako==1.3.10
MarkupSafe 
matplotlib-inline 
mkl-service==2.4.0
mkl_fft 
mkl_random 
more-itertools==10.7.0
mpmath 
mypy_extensions==1.1.0
narwhals 
nest_asyncio 
networkx 
numexpr 
numpy 
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==2.8.1
opentelemetry-api 
packaging 
pandas 
parso 
pathspec==0.12.1
pexpect 
pickleshare 
pillow 
platformdirs 
prompt_toolkit 
protobuf 
psutil 
ptyprocess 
pure_eval 
pyarrow 
pycparser 
pydantic==2.11.7
pydantic_core==2.33.2
Pygments 
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PySocks 
python-dateutil 
pytz 
PyYAML 
pyzmq 
redis==7.1.0
referencing 
regex==2024.11.6
requests==2.32.4
rpds-py 
safetensors==0.5.3
sentencepiece==0.2.0
six 
smmap 
sniffio==1.3.1
SQLAlchemy==2.0.41
stack_data 
streamlit 
sympy 
tenacity 
tokenizers==0.21.1
toml 
tomli==2.2.1
torch==2.7.1
tornado 
tqdm==4.67.1
traitlets 
transformers==4.52.4
triton==3.3.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata 
urllib3 
watchdog 
wcwidth 
Werkzeug==3.1.3
wrapt 
WTForms==3.2.1
xformers==0.0.31
zipp 
